Update README.md
This was my first finished HTML/CSS webpage

This project was a great learning experience for me to learn positioning and the use of containers and divs.

Before this project I was a bit confused with the creation of responsive webites and how it would function on multiple devices. But I learned how to make that happen with the use of different programs or just simple alignments. 

Hope you enjoy!